rewrite transactions reducer to slice
